+++ daemon-rs/crates/daemon/src/platform/firewall/netlink/mod.rs
@@ -13,4 +13,5 @@ pub(crate) use chain::*;
 pub(crate) use rule::NftRule;
 pub(crate) use table::*;
 pub(crate) use types::*;
+#[allow(unused_imports)]
 pub(crate) use zone::*;